+++ invai-back/src/test/java/es/caib/invai/back/service/mapper/maintenance/general/systemType/SystemTypeMapperTest.java
@@ -111,6 +111,6 @@ class SystemTypeMapperTest {
 
     @Test
     void map_nullId_returnsNull() {
-        assertNull(mapper.map((Long) null));
+        assertNull(mapper.map(null));
     }
 }
