+++ invai-back/src/test/java/es/caib/invai/back/service/mapper/maintenance/responsible/authorizationType/AuthorizationTypeMapperTest.java
@@ -107,6 +107,6 @@ class AuthorizationTypeMapperTest {
 
     @Test
     void map_nullId_returnsNull() {
-        assertNull(mapper.map((Long) null));
+        assertNull(mapper.map(null));
     }
 }
